+++ src/main/java/insane96mcp/survivalreimagined/data/generator/SRItemTagsProvider.java
@@ -63,6 +63,8 @@ public class SRItemTagsProvider extends ItemTagsProvider {
                 .add(Keego.HELMET.get(), Keego.CHESTPLATE.get(), Keego.LEGGINGS.get(), Keego.BOOTS.get());
 
         tag(ItemTags.ARROWS).add(Fletching.QUARTZ_ARROW_ITEM.get(), Fletching.DIAMOND_ARROW_ITEM.get(), Fletching.EXPLOSIVE_ARROW_ITEM.get(), Fletching.TORCH_ARROW_ITEM.get());
+
+        tag(ItemTags.BEACON_PAYMENT_ITEMS).add(Items.NETHER_STAR).add(Durium.INGOT.get(), SoulSteel.INGOT.get(), Keego.GEM.get());
         //Mod's
         tag(Hoes.DISABLED_HOES)
                 .add(Items.WOODEN_HOE);